feat(components): separate select logic into several files
